This is Me
As a scholar with a fervent passion for technology, I am embarking on a journey to major in Computer Science and Business at Brandeis University. My commitment to academic excellence is complemented by an unwavering enthusiasm for programming, web development, and data science. I thrive on challenges and am eager to contribute my skills to innovative projects.